+++ ImageHosting.Storage/Features/Images/Handlers/UploadFileHandler.cs
@@ -3,18 +3,25 @@ using System.Threading;
 using System.Threading.Tasks;
 using ImageHosting.Persistence.ValueTypes;
 using ImageHosting.Storage.Features.Images.Models;
+using ImageHosting.Storage.Features.Images.Services;
 using ImageHosting.Storage.Generic;
 using Microsoft.AspNetCore.Http;
 
-namespace ImageHosting.Storage.Features.Images.Services;
+namespace ImageHosting.Storage.Features.Images.Handlers;
 
-public class UploadFileService(
+public interface IUploadFileHandler
+{
+    Task<ImageUploadedResponse> UploadAsync(UserId userId, ImageId imageId, IFormFile formFile, bool hidden, DateTime uploadedAt,
+        CancellationToken cancellationToken = default);
+}
+
+public class UploadFileHandler(
     IFileUploadCommandFactory fileUploadCommandFactory,
     IMetadataUploadCommandFactory metadataUploadCommandFactory,
     IPublishNewMessageCommandFactory publishNewMessageCommandFactory)
-    : IUploadFileService
+    : IUploadFileHandler
 {
-    public async Task<ReadImageResponse> UploadAsync(UserId userId, ImageId imageId, IFormFile formFile, bool hidden,
+    public async Task<ImageUploadedResponse> UploadAsync(UserId userId, ImageId imageId, IFormFile formFile, bool hidden,
         DateTime uploadedAt, CancellationToken cancellationToken = default)
     {
         var fileUploadCommand = fileUploadCommandFactory.CreateCommand(userId, imageId, formFile);
@@ -29,6 +36,6 @@ public class UploadFileService(
 
         await commands.ExecuteAsync(cancellationToken).ConfigureAwait(false);
 
-        return new ReadImageResponse(imageId, userId, formFile.FileName, hidden, uploadedAt);
+        return new ImageUploadedResponse(imageId, userId, formFile.FileName, hidden, uploadedAt);
     }
 }
